pyproject.toml
[tool.poetry]
name = "hgraph-platform-tools"
version = "0.1.0"
description = ""
authors = ["Simon Kelly <si.j.kelly@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pandas = "*"
numpy = "*"
pymongo = "*"
hgraph = "*"
jinja2 = "*"
requests = "*"
kafka-python = "*"
xmlschema = "*"
python-dotenv = "*"

[tool.poetry.scripts]
hgraph = "cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
